rts: Split up rts/include/stg/MachRegs.h by arch
(cherry picked from commit b9c85181ee773aac63d88fc3c2169601382bfabf)
